requirements.txt
alabaster==0.7.9
appdirs==1.4.0
astroid==1.4.9
awsebcli==3.7.8
Babel==2.3.4
bcrypt==3.1.2
beautifulsoup4==4.5.1
billiard==3.5.0.2
bioservices==1.4.12
bleach==1.5.0
blessed==1.9.5
botocore==1.4.56
Bottleneck==1.1.0
cement==2.8.2
cffi==1.9.1
chardet==2.3.0
colorama==0.3.7
coverage==4.3.1
cryptography==1.7.1
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9
django-allauth==0.29.0
django-appconf==1.0.2
django-autocomplete-light==3.1.8
django-compressor==2.1
django-debug-toolbar==1.5
django-devserver==0.8.0
django-extensions==1.7.5
django-grappelli==2.8.1
django-grappelli-autocomplete-fk-edit-link==0.1.1
django-hstore==1.4.2
django-nose==1.4.4
django-simple-history==1.8.1
django-super-inlines==0.1.4
django-tables2==1.2.9
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
easydev==0.9.25
Fabric==1.13.1
Genesis-PyAPI==1.2.0
gevent==1.1.2
greenlet==0.4.10
grequests==0.3.0
gunicorn==19.6.0
html5lib==0.9999999
idna==2.2
imagesize==0.7.1
isort==4.2.5
Jinja2==2.8.1
jmespath==0.9.0
joblib==0.10.2
lazy-object-proxy==1.2.2
line-profiler==1.0
lxml==3.6.1
MarkupSafe==0.23
mccabe==0.5.3
mock==2.0.0
nose==1.3.7
numpy==1.11.1
oauthlib==2.0.1
pandas==0.18.1
paramiko==2.1.1
pathspec==0.3.4
pbr==1.10.0
pep8==1.7.0
pexpect==4.2.0
Pillow==3.4.2
plotly==1.12.9
prettytable==0.7.2
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
pyflakes==1.3.0
Pygments==2.1.3
pylint==1.6.4
pyqtgraph==0.9.10
python-dateutil==2.5.3
python-memcached==1.58
python3-openid==3.0.10
pytz==2016.6.1
PyYAML==3.12
rcssmin==1.0.6
requests==2.9.1
requests-cache==0.4.12
requests-oauthlib==0.7.0
rjsmin==1.0.12
scikit-learn==0.17.1
scipy==0.18.0
semantic-version==2.5.0
six==1.10.0
slumber==0.7.1
snowballstemmer==1.2.1
Sphinx==1.5.1
sqlparse==0.2.1
suds-jurko==0.6
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
wrapt==1.10.8
xlrd==1.0.0
xmltodict==0.10.2
